+++ CHANGELOG.md
@@ -6,12 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Added
 
+- Plot Gaussian process from hyperparameter Bayesian optimization.
 - Plot cluster losses and population histogram using matplotlib.
 - Option to use a sequential reduction optimizer for Bayesian optimization.
 - Specify Gaussian process keyword arguments for the final iterative training task.
 
 ### Changed
 
+- Default ``gp_params`` for Bayesian optimization.
 - MD5 hashes are no longer stored in bytes.
 - Do not include training set in any problematic clustering.
 Training structures are not included in dataset clustering or plots.